Inject `spec.collection.type` for cluster-logging channels >= 5.5
We add some logic to parse the provided channel to determine whether
we're on a 5.5+ channel. The parsing logic supports arbitrary channel
names but expects the format to be `<channel>-<major>.<minor>`. If
parsing fails, the logic prints a warning and assumes that the channel
is >= 5.5.

We need to make sure we only add the field `spec.collector.type` if
we're on at least 5.5, because earlier versions of the cluster-logging
operator will remove the field `spec.collection.type` from the CR which
will cause a loop between ArgoCD adding the field and the operator
removing it.

This is required to stop the collector pods from constantly restarting,
cf. https://access.redhat.com/solutions/6976455

Fixes #68