Switch fetching of fluentd alert YAML file to lookup map
We need to specify different locations of the fluentd alerts YAML file
in the upstream repository based on the logging stack release, as the
file got moved to `files/collector/fluentd_prometheus_alerts.yaml` for
release 5.5.

We implement this as a lookup map in the component class. Because we
don't have a way to implement a fallback value for keys which aren't
present in the lookup map, this change removes out-of-the-box support
for future logging stack versions.