Refactor version handling for cluster-logging
We introduce a new parameter `version` which is used to define default
values for parameters `channel` and `alerts`. This change is
backwards-compatible, as users can still override parameters `channel`
and `alerts`.

Because this removes the commit ID based setup for fetching the alert
rules YAML, we also remove the renovate regex manager config to
automatically generate PRs for new alert commit IDs.