Adjust default rate limits
The component configures fluentd as the default log forwarder. The
default chunk size limit in fluentd is 8 MB which the burst size limit
of 9 MB accounts for.

The ingestion rate defines the MB/s limit for a tenant. OpenShift
Logging uses the following three tenants:
* application
* audit
* infrastructure

The max allowed volume for a tenant per day can be calculated with:
`ingestionRate * 60 * 60 * 24 / 1024 = GB logs`

The default of 5 MB/s allows up to ~420 GB of logs per day for a tenant.