Fix console plugin configuration for Loki
The current Patch would work with strategy
`application/merge-patch+json`. However, this actually replaces the list
and does not append the item. If a second Patch would also update the
plugins list, both Patches would continuously replace the plugins list
with their desired version.

Using strategy `application/json-patch+json` with the add operation
would endlessly append the item to the list. If the list would be
configured as `x-kubernetes-list-type: set` in the CRD, the add
operation would fail after the initial run.

Given these circumstances, patching the parameters of
component-openshift4-console is the way to go for now.