Fix openshift-console config when deploying EFK stack
Previously, the component set parameter `openshift4_console` to `null`
when the EFK stack is being deployed, this breaks catalog compilation
for existing setups because it wipes the openshift4-console component
defaults, if this component's defaults are included after the console
defaults.

This commit fixes the bug by setting `openshift4_console: {}` if the
EFK stack is being deployed.