Bump replica count for LokiStack read path to 2 for 1x.demo
Cluster-logging 5.8 introduces PDBs for the Loki stack. Unfortunately
the PDBs are created unconditionally, and block node drains when
deploying LokiStack with `1x.demo`.

We adjust the replica count for the LokiStack read path (gateway,
querier, queryfrontend) to 2, so that we can drain infra nodes without
violating the PDBs for those components.